+++ eth2/proto_array_fork_choice/src/bin.rs
@@ -1,6 +1,8 @@
+#[allow(dead_code)]
 mod error;
 mod fork_choice_test_definition;
 mod proto_array;
+#[allow(dead_code)]
 mod proto_array_fork_choice;
 mod ssz_container;
