+++ eth2/proto_array_fork_choice/src/lib.rs
@@ -1,8 +1,7 @@
 mod error;
-pub mod fork_choice_test_definition;
 mod proto_array;
 mod proto_array_fork_choice;
 mod ssz_container;
 
+pub use crate::proto_array_fork_choice::ProtoArrayForkChoice;
 pub use error::Error;
-pub use proto_array_fork_choice::ProtoArrayForkChoice;
